refactor: extract MultiProviderHookExecutor and HookExecutionContext from MultiProvider
